Initial Commit
Application that handles the gathering, processing and storing of meteo data for specified locations. It works with MongoDB Atlas as a database, and it's written mostly in python. The goal is to have it prepare data for many APIs to work on.

Right now it works by running the main.py file.
It does logs each time its run and it stores on daily log files.